Refactoring package organization
