+++ app/src/main/java/br/com/luisfernandez/github/client/feature/issuelist/IssueListRepository.kt
@@ -1,8 +1,8 @@
 package br.com.luisfernandez.github.client.feature.issuelist
 
-import br.com.luisfernandez.github.client.http.model.GitHubErrorBody
+import br.com.luisfernandez.github.client.mvvm.repository.http.pojo.GitHubErrorBody
 import br.com.luisfernandez.github.client.mvvm.repository.ResultWrapper
-import br.com.luisfernandez.github.client.pojo.IssueResponse
+import br.com.luisfernandez.github.client.mvvm.repository.pojo.IssueResponse
 
 interface IssueListRepository {
     suspend fun loadIssueList(owner: String, repoName: String): ResultWrapper<List<IssueResponse>, GitHubErrorBody>
