+++ app/src/main/java/br/com/luisfernandez/github/client/feature/issuelist/IssueListRepositoryImpl.kt
@@ -1,10 +1,10 @@
 package br.com.luisfernandez.github.client.feature.issuelist
 
-import br.com.luisfernandez.github.client.api.GitHubService
-import br.com.luisfernandez.github.client.http.model.GitHubErrorBody
-import br.com.luisfernandez.github.client.http.v2.jw.DeferredResponseHandler
+import br.com.luisfernandez.github.client.mvvm.repository.api.GitHubService
+import br.com.luisfernandez.github.client.mvvm.repository.http.pojo.GitHubErrorBody
+import br.com.luisfernandez.github.client.mvvm.repository.http.DeferredResponseHandler
 import br.com.luisfernandez.github.client.mvvm.repository.ResultWrapper
-import br.com.luisfernandez.github.client.pojo.IssueResponse
+import br.com.luisfernandez.github.client.mvvm.repository.pojo.IssueResponse
 
 class IssueListRepositoryImpl(
         private val gitHubService: GitHubService
