+++ app/src/main/java/br/com/luisfernandez/github/client/feature/pullrequest/PullRequestListView.kt
@@ -1,8 +1,8 @@
 package br.com.luisfernandez.github.client.feature.pullrequest
 
-import br.com.luisfernandez.github.client.http.model.GitHubErrorBody
-import br.com.luisfernandez.github.client.mvp.LoadContentView
-import br.com.luisfernandez.github.client.pojo.PullRequestResponse
+import br.com.luisfernandez.github.client.mvvm.repository.http.pojo.GitHubErrorBody
+import br.com.luisfernandez.github.client.mvvm.view.LoadContentView
+import br.com.luisfernandez.github.client.mvvm.repository.pojo.PullRequestResponse
 
 
 /**
