+++ app/src/main/java/br/com/luisfernandez/github/client/feature/pullrequest/PullRequestModelImpl.kt
@@ -1,6 +1,6 @@
 package br.com.luisfernandez.github.client.feature.pullrequest
 
-import br.com.luisfernandez.github.client.api.GitHubService
+import br.com.luisfernandez.github.client.mvvm.repository.api.GitHubService
 
 class PullRequestModelImpl (
     private val gitHubService: GitHubService
