+++ app/src/main/java/br/com/luisfernandez/github/client/feature/repolist/RepoListModelImpl.kt
@@ -1,8 +1,8 @@
 package br.com.luisfernandez.github.client.feature.repolist
 
-import br.com.luisfernandez.github.client.http.model.GitHubErrorBody
+import br.com.luisfernandez.github.client.mvvm.repository.http.pojo.GitHubErrorBody
 import br.com.luisfernandez.github.client.mvvm.repository.ResultWrapper
-import br.com.luisfernandez.github.client.pojo.Repo
+import br.com.luisfernandez.github.client.mvvm.repository.pojo.Repo
 
 class RepoListModelImpl (
     private val repoListRepository: RepoListRepository
