+++ app/src/main/java/br/com/luisfernandez/github/client/feature/repolist/RepoListRepositoryImpl.kt
@@ -1,12 +1,12 @@
 package br.com.luisfernandez.github.client.feature.repolist
 
-import br.com.luisfernandez.github.client.api.GitHubService
+import br.com.luisfernandez.github.client.mvvm.repository.api.GitHubService
 import br.com.luisfernandez.github.client.mvvm.repository.ResultWrapper
 import br.com.luisfernandez.github.client.mvvm.repository.ResultWrapperAdapter
-import br.com.luisfernandez.github.client.http.model.GitHubErrorBody
-import br.com.luisfernandez.github.client.http.v2.jw.DeferredResponseHandler
-import br.com.luisfernandez.github.client.pojo.Repo
-import br.com.luisfernandez.github.client.pojo.RepoListResponse
+import br.com.luisfernandez.github.client.mvvm.repository.http.pojo.GitHubErrorBody
+import br.com.luisfernandez.github.client.mvvm.repository.http.DeferredResponseHandler
+import br.com.luisfernandez.github.client.mvvm.repository.pojo.Repo
+import br.com.luisfernandez.github.client.mvvm.repository.pojo.RepoListResponse
 
 class RepoListRepositoryImpl(
         private val gitHubService: GitHubService
