+++ app/src/main/java/br/com/luisfernandez/github/client/koin/RepositoryModule.kt
@@ -1,6 +1,6 @@
 package br.com.luisfernandez.github.client.koin
 
-import br.com.luisfernandez.github.client.api.GitHubService
+import br.com.luisfernandez.github.client.mvvm.repository.api.GitHubService
 import br.com.luisfernandez.github.client.feature.issuelist.IssueListRepository
 import br.com.luisfernandez.github.client.feature.issuelist.IssueListRepositoryImpl
 import br.com.luisfernandez.github.client.feature.pullrequest.PullRequestRepository
