+++ app/src/main/java/br/com/luisfernandez/github/client/mvvm/repository/http/pojo/GitHubErrorBody.kt
@@ -1,4 +1,4 @@
-package br.com.luisfernandez.github.client.http.model
+package br.com.luisfernandez.github.client.mvvm.repository.http.pojo
 
 import com.google.gson.annotations.SerializedName
