+++ app/src/main/java/br/com/luisfernandez/github/client/mvvm/repository/pojo/RepoListResponse.kt
@@ -1,4 +1,4 @@
-package br.com.luisfernandez.github.client.pojo
+package br.com.luisfernandez.github.client.mvvm.repository.pojo
 
 import com.google.gson.annotations.SerializedName
