+++ app/src/main/java/br/com/luisfernandez/github/client/mvvm/view/ErrorHandlerView.kt
@@ -1,6 +1,6 @@
-package br.com.luisfernandez.github.client.mvp
+package br.com.luisfernandez.github.client.mvvm.view
 
-import br.com.luisfernandez.github.client.http.model.ServerError
+import br.com.luisfernandez.github.client.mvvm.repository.http.pojo.ServerError
 
 /**
  * Created by luisfernandez on 12/05/18.
