+++ app/src/main/java/br/com/luisfernandez/github/client/mvvm/view/event/OnNeedUpdateEvent.kt
@@ -1,3 +1,3 @@
-package br.com.luisfernandez.github.client.mvvm.event
+package br.com.luisfernandez.github.client.mvvm.view.event
 
 class OnNeedUpdateEvent<EventPayload> : ViewEvent<EventPayload>()
